requirements.txt
absl-py==2.1.0
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
Bottleneck @ file:///croot/bottleneck_1707864210935/work
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools==5.5.0
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
clarabel==0.9.0
comm @ file:///croot/comm_1709322850197/work
contourpy @ file:///opt/conda/conda-bld/contourpy_1663827406301/work
cvxpy==1.5.2
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
ecos==2.0.14
exceptiongroup==1.2.2
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
filelock @ file:///croot/filelock_1700591183607/work
fonttools @ file:///croot/fonttools_1713551344105/work
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455532332/work
google-auth==2.34.0
google-auth-oauthlib==1.0.0
grpcio==1.66.1
idna @ file:///croot/idna_1714398848350/work
importlib-metadata @ file:///croot/importlib_metadata-suite_1704813515092/work
importlib_resources @ file:///croot/importlib_resources-suite_1720641103994/work
iniconfig==2.0.0
ipykernel @ file:///croot/ipykernel_1705933831282/work
ipython @ file:///croot/ipython_1691532092695/work
jedi @ file:///croot/jedi_1721058342488/work
Jinja2 @ file:///croot/jinja2_1716993405101/work
joblib==1.4.2
jupyter_client @ file:///croot/jupyter_client_1699455897726/work
jupyter_core @ file:///croot/jupyter_core_1718818295206/work
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
Markdown==3.7
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib @ file:///croot/matplotlib-suite_1693812469450/work
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
networkx @ file:///croot/networkx_1690561992265/work
numexpr @ file:///croot/numexpr_1683221822650/work
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
oauthlib==3.2.2
osqp==0.6.7.post1
packaging @ file:///croot/packaging_1720101850331/work
pandas @ file:///croot/pandas_1692289311655/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
pillow @ file:///croot/pillow_1721059439630/work
platformdirs @ file:///croot/platformdirs_1692205439124/work
pluggy==1.5.0
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
protobuf==5.28.1
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments @ file:///croot/pygments_1684279966437/work
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
pytest==8.3.2
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
pytz @ file:///croot/pytz_1713974312559/work
PyYAML @ file:///croot/pyyaml_1698096049011/work
pyzmq @ file:///croot/pyzmq_1705605076900/work
qdldl==0.1.7.post4
requests @ file:///croot/requests_1721410876868/work
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
scs==3.2.7
seaborn @ file:///croot/seaborn_1718302919398/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
sympy @ file:///croot/sympy_1724938189289/work
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tomli==2.0.1
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tornado @ file:///croot/tornado_1718740109488/work
tqdm==4.66.5
traitlets @ file:///croot/traitlets_1718227057033/work
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
unicodedata2 @ file:///croot/unicodedata2_1713212950228/work
urllib3 @ file:///croot/urllib3_1718912636303/work
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
Werkzeug==3.0.4
zeus-mcmc==2.5.4
zipp @ file:///croot/zipp_1704206909481/work
